Ensure Package Compatibility Of Audio Search Subsystem

+++ src/main/java/org/hust/audioSearch/shazam/Peak.java
@@ -2,9 +2,9 @@ package org.hust.audioSearch.shazam;
 
 public class Peak {
     public short frequency;
-    public int time;
+    public double time;
 
-    public Peak(short frequency, int time){
+    public Peak(short frequency, double time){
         this.frequency = frequency;
         this.time = time;
     }
